# internal/static/swagger.json
{
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "The purpose of this service is to provide an application to help manage db switch task",
    "title": "DCSwitch API",
    "version": "0.0.1"
  },
  "host": "127.0.0.1",
  "paths": {
...
    "/switch/detail/editComment/{id}": {
      "get": {
        "description": "编辑准备项",
        "operationId": "EditSwitchDetailParam",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Id",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "comment": {
                  "type": "string",
                  "x-go-name": "Comment"
                },
                "endTime": {
                  "type": "string",
                  "x-go-name": "EndTime"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/switch/detail/editEndTIme/{id}": {
      "get": {
        "description": "编辑准备项",
        "operationId": "EditSwitchDetailParam",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Id",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "comment": {
                  "type": "string",
                  "x-go-name": "Comment"
                },
                "endTime": {
                  "type": "string",
                  "x-go-name": "EndTime"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/switch/details/{versionId}": {
      "get": {
        "description": "获取所有切换版本信息",
        "operationId": "GetAllDetailsParam",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "VersionId",
            "name": "versionId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetAllDetailsResp"
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/switch/domain/{id}": {
      "get": {
        "description": "获取所有切换版本信息",
        "operationId": "GetSwitchDomainParam",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetSwitchDomainResp"
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/switch/failed/details/{moduleId}": {
      "get": {
        "description": "获取所有切换版本信息",
        "operationId": "GetModuleDetailsParam",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ModuleId",
            "name": "moduleId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetModuleDetailsResp"
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/switch/failed/details/{versionId}": {
      "get": {
        "description": "获取所有切换版本信息",
        "operationId": "GetModuleDetailsParam",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ModuleId",
            "name": "moduleId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetModuleDetailsResp"
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/switch/module/domain/{moduleId}": {
      "get": {
        "description": "获取所有切换版本信息",
        "operationId": "GetModuleParam",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ModuleId",
            "name": "moduleId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetModuleResp"
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/switch/module/progress/{moduleId}": {
      "get": {
        "description": "获取所有切换版本信息",
        "operationId": "GetModuleProgressParam",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ModuleId",
            "name": "moduleId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetModuleProgressResp"
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/switch/modules/domain/{domainId}": {
      "get": {
        "description": "获取所有切换版本信息",
        "operationId": "GetModulesParam",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "DomainId",
            "name": "domainId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetModulesResp"
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/switch/modules/{versionId}": {
      "get": {
        "description": "获取所有切换版本信息",
        "operationId": "GetAllModulesParam",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "VersionId",
            "name": "versionId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetAllModulesResp"
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/switch/preparation/edit/{id}": {
      "get": {
        "description": "编辑准备项",
        "operationId": "EditPrepareParam",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Id",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string",
                  "x-go-name": "Content"
                },
                "title": {
                  "type": "string",
                  "x-go-name": "Title"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/switch/preparation/{id}": {
      "get": {
        "description": "获取所有切换版本信息",
        "operationId": "GetPrepareParam",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetPrepareResp"
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/switch/preparations/{versionId}": {
      "get": {
        "description": "获取所有切换版本信息",
        "operationId": "GetAllPreparesParam",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "VersionId",
            "name": "versionId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetAllPreparesResp"
          },
          "500": {
            "description": ""
          }
        }
      }
    },
...
    "/switch/version/progress/{versionId}": {
      "get": {
        "description": "获取所有切换版本信息",
        "operationId": "GetVersionProgressParam",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "VersionId",
            "name": "versionId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetVersionProgressResp"
          },
          "500": {
            "description": ""
          }
        }
      }
    },
...
  },
  "definitions": {
    "Detail": {
      "type": "object",
      "properties": {
        "BeginTime": {
          "type": "string"
        },
        "Comment": {
          "type": "string"
        },
        "Detail": {
          "type": "string"
        },
        "Duration": {
          "type": "string"
        },
        "EndTime": {
          "type": "string"
        },
        "Id": {
          "type": "integer",
          "format": "int64"
        },
        "Module": {
          "$ref": "#/definitions/Module"
        },
        "Status": {
          "type": "integer",
          "format": "int64"
        }
      },
      "x-go-package": "dcswitch/internal/domain"
    },
    "DetailExcel": {
      "type": "object",
      "properties": {
        "Detail": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "x-go-package": "dcswitch/internal/domain"
    },
    "Domain": {
      "description": "version_domain 切换域",
      "type": "object",
      "properties": {
        "AppDomain": {
          "type": "string"
        },
        "DelFlag": {
          "type": "integer",
          "format": "int64"
        },
        "DomainArchitect": {
          "type": "string"
        },
        "Id": {
          "type": "integer",
          "format": "int64"
        },
        "Type": {
          "type": "string"
        },
        "Version": {
          "$ref": "#/definitions/SwitchVersion"
        }
      },
      "x-go-package": "dcswitch/internal/domain"
    },
    "Module": {
      "type": "object",
      "properties": {
        "Details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DetailExcel"
          }
        },
        "File": {
          "type": "string"
        },
        "Id": {
          "type": "integer",
          "format": "int64"
        },
        "Module": {
          "type": "string"
        }
      },
      "x-go-package": "dcswitch/internal/domain"
    },
    "ModuleProgress": {
      "type": "object",
      "properties": {
        "BeginTime": {
          "type": "string"
        },
        "Duration": {
          "type": "string"
        },
        "EndTime": {
          "type": "string"
        },
        "Module": {
          "type": "string"
        },
        "ModuleStatus": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ModuleStatus"
          }
        }
      },
      "x-go-package": "dcswitch/internal/domain"
    },
    "ModuleStatus": {
      "type": "object",
      "properties": {
        "Count": {
          "type": "integer",
          "format": "int64"
        },
        "Status": {
          "type": "integer",
          "format": "int64"
        }
      },
      "x-go-package": "dcswitch/internal/domain"
    },
    "Prepare": {
      "description": "version_domain_prepare 切换准备项",
      "type": "object",
      "properties": {
        "Content": {
          "type": "string"
        },
        "CreateTime": {
          "type": "string"
        },
        "CreateUser": {
          "type": "string"
        },
        "DelFlag": {
          "type": "integer",
          "format": "int64"
        },
        "Domain": {
          "$ref": "#/definitions/Domain"
        },
        "Id": {
          "type": "integer",
          "format": "int64"
        },
        "Sequence": {
          "type": "integer",
          "format": "int64"
        },
        "Status": {
          "type": "integer",
          "format": "int64"
        },
        "Title": {
          "type": "string"
        },
        "UpdateTime": {
          "type": "string"
        },
        "UpdateUser": {
          "type": "string"
        }
      },
      "x-go-package": "dcswitch/internal/domain"
    },
    "SwitchDomain": {
      "description": "SwitchDomain 切换域",
      "type": "object",
      "properties": {
        "AppDomain": {
          "type": "string"
        },
        "Id": {
          "type": "integer",
          "format": "int64"
        }
      },
      "x-go-package": "dcswitch/internal/domain"
    },
    "SwitchVersion": {
      "description": "SwitchVersion 切换演练版本",
      "type": "object",
      "properties": {
        "CreateTime": {
          "type": "string"
        },
        "CreateUser": {
          "type": "string"
        },
        "Id": {
          "type": "integer",
          "format": "int64"
        },
        "VersionDate": {
          "type": "string"
        }
      },
      "x-go-package": "dcswitch/internal/domain"
    }
  },
  "responses": {
    "GetAllDetailsResp": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Detail"
            },
            "x-go-name": "Details"
          }
        }
      }
    },
    "GetAllModulesResp": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Module"
            },
            "x-go-name": "Modules"
          }
        }
      }
    },
    "GetAllPreparesResp": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "prepares": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Prepare"
            },
            "x-go-name": "Prepares"
          }
        }
      }
    },
...
    "GetModuleDetailsResp": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Detail"
            },
            "x-go-name": "Details"
          }
        }
      }
    },
    "GetModuleProgressResp": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "progress": {
            "$ref": "#/definitions/ModuleProgress"
          }
        }
      }
    },
    "GetModuleResp": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "module": {
            "$ref": "#/definitions/Module"
          }
        }
      }
    },
...
    "GetModulesResp": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Module"
            },
            "x-go-name": "Modules"
          }
        }
      }
    },
    "GetPrepareResp": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "prepare": {
            "$ref": "#/definitions/Prepare"
          }
        }
      }
    },
    "GetSwitchDomainResp": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "switchDomain": {
            "$ref": "#/definitions/SwitchDomain"
          }
        }
      }
    },
    "GetVersionProgressResp": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "progress": {
            "$ref": "#/definitions/ModuleProgress"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "TOKEN",
      "in": "header"
    }
  },
  "security": [
    {
      "api_key": []
    }
  ]
}